requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
distlib==0.3.9
Django==4.2.20
django-bootstrap-v5==1.0.11
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.18.0
inflection==0.5.1
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.1
uritemplate==4.1.1
virtualenv==20.29.3
